new version i did over the weekend.
functional game with a functioning double and split function. however i did use a decent amount of ai during the process, so this should not be used for the final version since i cant really explain why most of the JS works the way it does. it can be used for good inspiration but again dont copy and paste any of the code unless you can understand and explain what it does. the html and css could just be copied over but might not be entirely necessary. we'll see. hope i did the branch thing right but not entirly sure of it so im not deleting any additional files yet since im not sure if i would remove it on the main branch or not.